fixes building using ndk 14 (android 2.3 support)
this is a lazy fix, simply ifdef's out stuff that was missing in ndk14 with no attempt to work around it's limitations,however this is good enough for my usage.

the impl.h was missing includes for both stdint.h and imgui.h. i decided to change the return type from int32_t to int (same thing on android, no chance int being 16bit).